added rate this book text in rating component
