+++ src/app/books/[olid]/components/BookCover.tsx
@@ -1,35 +1,14 @@
 'use client'
 import useBookDetails from '@/app/hooks/useBookDetails'
-import {
-  DiscordLogoIcon,
-  ExternalLinkIcon,
-  InstagramLogoIcon,
-  LinkedInLogoIcon,
-  TwitterLogoIcon,
-} from '@radix-ui/react-icons'
-import {
-  Box,
-  Button,
-  Card,
-  Dialog,
-  Flex,
-  IconButton,
-  Text,
-} from '@radix-ui/themes'
+import { ExternalLinkIcon } from '@radix-ui/react-icons'
+import { Box, Button, Card, Flex, Text } from '@radix-ui/themes'
 import React, { useState } from 'react'
 import { RotatingSquare } from 'react-loader-spinner'
 import NextLink from 'next/link'
-import { Rating } from 'react-simple-star-rating'
+import BookRating from './BookRating'
 
 export default function BookCover({ bookId }: { bookId: string }) {
   const { work, book, loading, error } = useBookDetails({ olid: bookId })
-  const [rating, setRating] = useState(0)
-  const [openDialog, setOpenDialog] = useState(false)
-
-  const handleRating = (rate: number) => {
-    setRating(rate)
-    setOpenDialog(true)
-  }
 
   return (
     <>
@@ -67,12 +46,7 @@ export default function BookCover({ bookId }: { bookId: string }) {
               />
             </picture>
             <Box data-testid='book-details-rating-container'>
-              <Rating
-                size={30}
-                SVGclassName='inline-block'
-                onClick={handleRating}
-                initialValue={rating}
-              />
+              <BookRating book={book} />
             </Box>
             <Box>
               <Button
@@ -102,49 +76,6 @@ export default function BookCover({ bookId }: { bookId: string }) {
           </Text>
         </Card>
       )}
-      <Dialog.Root
-        onOpenChange={setOpenDialog}
-        open={openDialog}
-        data-testid='book-details-appreciation-dialog'
-      >
-        <Dialog.Content>
-          <Dialog.Title>
-            Thanks for rating {book ? book.title : 'this book'}.
-          </Dialog.Title>
-          <Dialog.Description>
-            Your feedback matters. Thank you for helping us improve our
-            platform. Happy reading!
-          </Dialog.Description>
-          <Flex direction='row' gap='5' justify='center' align='center' p='5'>
-            <IconButton asChild size='3' radius='full' color='blue'>
-              <NextLink href='https://twitter.com' target='_blank'>
-                <TwitterLogoIcon width='28' height='28' />
-              </NextLink>
-            </IconButton>
-            <IconButton asChild size='3' radius='full' color='pink'>
-              <NextLink
-                href='https://www.instagram.com/spenz_dev'
-                target='_blank'
-              >
-                <InstagramLogoIcon width='28' height='28' />
-              </NextLink>
-            </IconButton>
-            <IconButton asChild size='3' radius='full' color='indigo'>
-              <NextLink
-                href='https://www.linkedin.com/in/spencer-rafada/'
-                target='_blank'
-              >
-                <LinkedInLogoIcon width='28' height='28' />
-              </NextLink>
-            </IconButton>
-            <IconButton asChild size='3' radius='full' color='purple'>
-              <NextLink href='https://discord.com' target='_blank'>
-                <DiscordLogoIcon width='28' height='28' />
-              </NextLink>
-            </IconButton>
-          </Flex>
-        </Dialog.Content>
-      </Dialog.Root>
     </>
   )
 }
